Extracted the primary editor button (publish/update) to its own component
