+++ client/post-editor/editor-ground-control/index.jsx
@@ -23,6 +23,7 @@ const Card = require( 'components/card' ),
 	stats = require( 'lib/posts/stats' );
 
 import AsyncLoad from 'components/async-load';
+import EditorPublishButton from 'post-editor/editor-publish-button';
 
 export default React.createClass( {
 	displayName: 'EditorGroundControl',
@@ -33,7 +34,6 @@ export default React.createClass( {
 		isSaveBlocked: React.PropTypes.bool,
 		isPublishing: React.PropTypes.bool,
 		isSaving: React.PropTypes.bool,
-		onClose: React.PropTypes.func,
 		onPreview: React.PropTypes.func,
 		onPublish: React.PropTypes.func,
 		onSaveDraft: React.PropTypes.func,
@@ -43,7 +43,8 @@ export default React.createClass( {
 		site: React.PropTypes.object,
 		user: React.PropTypes.object,
 		userUtils: React.PropTypes.object,
-		type: React.PropTypes.string
+		type: React.PropTypes.string,
+		setDate: React.PropTypes.func
 	},
 
 	mixins: [ PureRenderMixin ],
@@ -55,7 +56,6 @@ export default React.createClass( {
 			isSaveBlocked: false,
 			isPublishing: false,
 			isSaving: false,
-			onClose: noop,
 			onPublish: noop,
 			onSaveDraft: noop,
 			post: null,
@@ -63,6 +63,7 @@ export default React.createClass( {
 			site: {},
 			user: null,
 			userUtils: null,
+			setDate: noop
 		};
 	},
 
@@ -139,65 +140,6 @@ export default React.createClass( {
 		}
 
 		return this.translate( 'Preview' );
-	},
-
-	trackPrimaryButton: function() {
-		const postEvents = {
-			update: 'Clicked Update Post Button',
-			schedule: 'Clicked Schedule Post Button',
-			requestReview: 'Clicked Request-Review Post Button',
-			publish: 'Clicked Publish Post Button',
-		};
-		const pageEvents = {
-			update: 'Clicked Update Page Button',
-			schedule: 'Clicked Schedule Page Button',
-			requestReview: 'Clicked Request-Review Page Button',
-			publish: 'Clicked Publish Page Button',
-		};
-		const buttonState = this.getPrimaryButtonState();
-		const eventString = postUtils.isPage( this.props.post ) ? pageEvents[ buttonState ] : postEvents[ buttonState ];
-		stats.recordEvent( eventString );
-		stats.recordEvent( 'Clicked Primary Button' );
-	},
-
-	getPrimaryButtonState: function() {
-		if (
-			postUtils.isPublished( this.props.savedPost ) &&
-			! postUtils.isBackDatedPublished( this.props.savedPost ) &&
-			! postUtils.isFutureDated( this.props.post ) ||
-			(
-				this.props.savedPost &&
-				this.props.savedPost.status === 'future' &&
-				postUtils.isFutureDated( this.props.post )
-			)
-		) {
-			return 'update';
-		}
-
-		if ( postUtils.isFutureDated( this.props.post ) ) {
-			return 'schedule';
-		}
-
-		if ( siteUtils.userCan( 'publish_posts', this.props.site ) ) {
-			return 'publish';
-		}
-
-		if ( this.props.savedPost && this.props.savedPost.status === 'pending' ) {
-			return 'update';
-		}
-
-		return 'requestReview';
-	},
-
-	getPrimaryButtonLabel: function() {
-		const primaryButtonState = this.getPrimaryButtonState();
-		const buttonLabels = {
-			update: this.translate( 'Update' ),
-			schedule: this.translate( 'Schedule' ),
-			publish: this.translate( 'Publish' ),
-			requestReview: this.translate( 'Submit for Review' ),
-		};
-		return buttonLabels[ primaryButtonState ];
 	},
 
 	getVerificationNoticeLabel: function() {
@@ -323,31 +265,8 @@ export default React.createClass( {
 			! this.props.isSaveBlocked;
 	},
 
-	isPrimaryButtonEnabled: function() {
-		return ! this.props.isPublishing &&
-			! this.props.isSaveBlocked &&
-			this.props.hasContent &&
-			! this.state.needsVerification;
-	},
-
 	toggleAdvancedStatus: function() {
 		this.setState( { showAdvanceStatus: ! this.state.showAdvanceStatus } );
-	},
-
-	onPrimaryButtonClick: function() {
-		this.trackPrimaryButton();
-
-		if ( postUtils.isPublished( this.props.savedPost ) &&
-			! postUtils.isBackDatedPublished( this.props.savedPost )
-		) {
-			return this.props.onSave();
-		}
-
-		if ( siteUtils.userCan( 'publish_posts', this.props.site ) ) {
-			return this.props.onPublish();
-		}
-
-		return this.props.onSave( 'pending' );
 	},
 
 	onSaveButtonClick: function() {
@@ -441,14 +360,10 @@ export default React.createClass( {
 						{ this.getPreviewLabel() }
 					</button>
 					<div className="editor-ground-control__publish-combo">
-						<button
-							className="editor-ground-control__publish-button button is-primary"
-							onClick={ this.onPrimaryButtonClick }
-							disabled={ ! this.isPrimaryButtonEnabled() }
+						<EditorPublishButton
+							{ ...this.props }
 							tabIndex={ 5 }
-						>
-							{ this.getPrimaryButtonLabel() }
-						</button>
+						/>
 						{ siteUtils.userCan( 'publish_posts', this.props.site ) &&
 							<button
 								ref="schedulePost"
